break up tests, add back lockfile
