+++ .gitignore
@@ -14,7 +14,6 @@ dist
 coverage/*
 .nyc_output/
 
-package-lock.json
 /yarn.lock
 /.eslintrc.json.tpl
 /.browserslistrc
