+++ .npmignore
@@ -15,7 +15,6 @@ Makefile
 *.old
 *.log
 tsconfig.json
-package-lock.json
 test
 .github
 .gitkeep
