+++ .npmrc
@@ -1,1 +1,1 @@
-package-lock=false
+package-lock=true
